back-end/requirements.txt
bcrypt
cryptography
fastapi
python-jose
python-multipart
SQLAlchemy
uvicorn
passlib
paho-mqtt
scikit-fuzzy
numpy
matplotlib
elementpath
httpx